Refactor: Add Attendees component
